+++ genericmoney/src/supportedNetworks.tsx
@@ -1,4 +1,4 @@
 import { InjectedConnector } from '@web3-react/injected-connector'
 export const injected = new InjectedConnector({
-  supportedChainIds: [97],
+  supportedChainIds: [97], // BSC
 })
