+++ genericmoney/src/utils.ts
@@ -1,0 +1,3 @@
+export function formatAddress(value: string, length: number = 4) {
+    return `${value.substring(0, length + 2)}...${value.substring(value.length - length)}`
+}
